+++ ptv_timetable/asyncapi.py
@@ -14,7 +14,7 @@ from . import _responsetypes
 __all__ = ["AsyncTimetableAPI", "METROPOLITAN_TRAIN", "METRO_TRAIN", "MET_TRAIN", "METRO", "TRAM", "BUS", "REGIONAL_TRAIN", "REG_TRAIN", "COACH", "VLINE", "EXPAND_ALL", "EXPAND_STOP", "EXPAND_ROUTE", "EXPAND_RUN", "EXPAND_DIRECTION", "EXPAND_DISRUPTION", "EXPAND_VEHICLE_DESCRIPTOR", "EXPAND_VEHICLE_POSITION", "EXPAND_NONE"]
 
 type ExpandType = Literal["All", "Stop", "Route", "Run", "Direction", "Disruption", "VehicleDescriptor", "VehiclePosition", "None"]
-type RouteType = Literal[0, 1, 2, 3]
+type RouteTypeType = Literal[0, 1, 2, 3]
 
 _logger: Final = logging.getLogger("ptv-timetable.ptv_timetable.asyncapi")
 """Logger for this module"""
@@ -177,7 +177,7 @@ class AsyncTimetableAPI(object):
 
         return [await Direction.aload(**item) for item in (await self.request(f"/v3/directions/route/{route_id}"))["directions"]]
 
-    async def get_direction(self: Self, direction_id: int, route_type: RouteType | None = None) -> list[Direction]:
+    async def get_direction(self: Self, direction_id: int, route_type: RouteTypeType | None = None) -> list[Direction]:
         """Returns the direction(s) of travel in the database with the specified identifier and route type. If ``route_type`` isn't specified, this will return directions of travel with the same identifier for all modes (which are likely unrelated to one another). Note that this returns a :class:`list` in both cases.
 
         If the direction is shared by multiple routes (e.g. Flinders Street), a :class:`~ptv_timetable.types.Direction` object will be added to the :class:`list` for *each* route.
@@ -193,7 +193,7 @@ class AsyncTimetableAPI(object):
 
     async def get_pattern(self: Self,
                           run_ref: str | int,
-                          route_type: RouteType,
+                          route_type: RouteTypeType,
                           stop_id: int | None = None,
                           date: datetime | str | None = None,
                           include_skipped_stops: bool | None = None,
@@ -251,7 +251,7 @@ class AsyncTimetableAPI(object):
         path = f"/v3/routes/{route_id}" + await self.generate_url_params(include_geopath=include_geopath, geopath_utc=geopath_date.astimezone(timezone.utc).isoformat())
         return await Route.aload(**(await self.request(path))["route"])
 
-    async def list_routes(self: Self, route_types: Iterable[RouteType] | RouteType | None = None, route_name: str | None = None) -> list[Route]:
+    async def list_routes(self: Self, route_types: Iterable[RouteTypeType] | RouteTypeType | None = None, route_name: str | None = None) -> list[Route]:
         """Returns all routes of all (or specified) types.
 
         :param route_types: Return only the routes of the specified type(s)
@@ -276,7 +276,7 @@ class AsyncTimetableAPI(object):
 
     async def get_run(self: Self,
                       run_ref: str | int,
-                      route_type: RouteType | None = None,
+                      route_type: RouteTypeType | None = None,
                       expand: Iterable[Literal["All", "VehicleDescriptor", "VehiclePosition", "None"]] | Literal["All", "VehicleDescriptor", "VehiclePosition", "None"] | None = None,
                       date: datetime | str | None = None,
                       include_geopath: bool | None = None
@@ -305,7 +305,7 @@ class AsyncTimetableAPI(object):
 
     async def list_runs(self: Self,
                         route_id: int,
-                        route_type: RouteType | None = None,
+                        route_type: RouteTypeType | None = None,
                         expand: Iterable[Literal["All", "VehicleDescriptor", "VehiclePosition", "None"]] | Literal["All", "VehicleDescriptor", "VehiclePosition", "None"] | None = None,
                         date: datetime | str | None = None
                         ) -> list[Run]:
@@ -333,7 +333,7 @@ class AsyncTimetableAPI(object):
     @overload
     async def get_stop(self: Self,
                        stop_id: int,
-                       route_type: RouteType,
+                       route_type: RouteTypeType,
                        stop_location: bool | None = None,
                        stop_amenities: bool | None = None,
                        stop_accessibility: bool | None = None,
@@ -348,7 +348,7 @@ class AsyncTimetableAPI(object):
     @overload
     async def get_stop(self: Self,
                        stop_id: str,
-                       route_type: RouteType,
+                       route_type: RouteTypeType,
                        stop_location: bool | None = None,
                        stop_amenities: bool | None = None,
                        stop_accessibility: bool | None = None,
@@ -364,7 +364,7 @@ class AsyncTimetableAPI(object):
     @overload
     async def get_stop(self: Self,
                        stop_id: str,
-                       route_type: RouteType,
+                       route_type: RouteTypeType,
                        stop_location: bool | None,
                        stop_amenities: bool | None,
                        stop_accessibility: bool | None,
@@ -378,7 +378,7 @@ class AsyncTimetableAPI(object):
 
     async def get_stop(self: Self,
                        stop_id: int | str,
-                       route_type: RouteType,
+                       route_type: RouteTypeType,
                        stop_location: bool | None = None,
                        stop_amenities: bool | None = None,
                        stop_accessibility: bool | None = None,
@@ -411,7 +411,7 @@ class AsyncTimetableAPI(object):
 
     async def list_stops(self: Self,
                          route_id: int,
-                         route_type: RouteType,
+                         route_type: RouteTypeType,
                          direction_id: int | None = None,
                          stop_disruptions: bool | None = None,
                          include_advertised_interchange: bool | None = None
@@ -438,7 +438,7 @@ class AsyncTimetableAPI(object):
     async def list_stops_near_location(self: Self,
                                        latitude: float,
                                        longitude: float,
-                                       route_types: Iterable[RouteType] | RouteType | None = None,
+                                       route_types: Iterable[RouteTypeType] | RouteTypeType | None = None,
                                        max_results: int | None = None,
                                        max_distance: float | None = None,
                                        stop_disruptions: bool | None = None
@@ -464,7 +464,7 @@ class AsyncTimetableAPI(object):
     # gtfs is not specified
     @overload
     async def list_departures(self: Self,
-                              route_type: RouteType,
+                              route_type: RouteTypeType,
                               stop_id: int,
                               route_id: int,
                               platform_numbers: None = None,
@@ -484,7 +484,7 @@ class AsyncTimetableAPI(object):
     # gtfs is not specified
     @overload
     async def list_departures(self: Self,
-                              route_type: RouteType,
+                              route_type: RouteTypeType,
                               stop_id: int,
                               route_id: None = None,
                               platform_numbers: Iterable[str | int] | str | int | None = None,
@@ -502,7 +502,7 @@ class AsyncTimetableAPI(object):
     # route_id is specified; gtfs is specified by keyword
     @overload
     async def list_departures(self: Self,
-                              route_type: RouteType,
+                              route_type: RouteTypeType,
                               stop_id: str,
                               route_id: int,
                               platform_numbers: None = None,
@@ -521,7 +521,7 @@ class AsyncTimetableAPI(object):
     # platform_numbers is specified; gtfs is specified by keyword
     @overload
     async def list_departures(self: Self,
-                              route_type: RouteType,
+                              route_type: RouteTypeType,
                               stop_id: str,
                               route_id: None = None,
                               platform_numbers: Iterable[str | int] | str | int | None = None,
@@ -540,7 +540,7 @@ class AsyncTimetableAPI(object):
     # route_id and gtfs are both specified by position
     @overload
     async def list_departures(self: Self,
-                              route_type: RouteType,
+                              route_type: RouteTypeType,
                               stop_id: str,
                               route_id: int,
                               platform_numbers: None,
@@ -559,7 +559,7 @@ class AsyncTimetableAPI(object):
     # also for case where both route_id and platform_numbers are not specified
     @overload
     async def list_departures(self: Self,
-                              route_type: RouteType,
+                              route_type: RouteTypeType,
                               stop_id: str,
                               route_id: None,
                               platform_numbers: Iterable[str | int] | str | int | None,
@@ -575,7 +575,7 @@ class AsyncTimetableAPI(object):
         ...
 
     async def list_departures(self: Self,
-                              route_type: RouteType,
+                              route_type: RouteTypeType,
                               stop_id: int | str,
                               route_id: int | None = None,
                               platform_numbers: Iterable[str | int] | str | int | None = None,
@@ -633,7 +633,7 @@ class AsyncTimetableAPI(object):
     async def list_disruptions(self: Self,
                                route_id: None = None,
                                stop_id: None = None,
-                               route_types: Iterable[RouteType] | RouteType | None = None,
+                               route_types: Iterable[RouteTypeType] | RouteTypeType | None = None,
                                disruption_modes: Iterable[Literal[1, 2, 3, 4, 5, 7, 8, 9, 10, 11, 12, 13, 14, 100]] | Literal[1, 2, 3, 4, 5, 7, 8, 9, 10, 11, 12, 13, 14, 100] | None = None,
                                disruption_status: Literal["current", "planned"] | None = None
                                ) -> list[Disruption]:
@@ -642,7 +642,7 @@ class AsyncTimetableAPI(object):
     async def list_disruptions(self: Self,
                                route_id: int | None = None,
                                stop_id: int | None = None,
-                               route_types: Iterable[RouteType] | RouteType | None = None,
+                               route_types: Iterable[RouteTypeType] | RouteTypeType | None = None,
                                disruption_modes: Iterable[Literal[1, 2, 3, 4, 5, 7, 8, 9, 10, 11, 12, 13, 14, 100]] | Literal[1, 2, 3, 4, 5, 7, 8, 9, 10, 11, 12, 13, 14, 100] | None = None,
                                disruption_status: Literal["current", "planned"] | None = None
                                ) -> list[Disruption]:
@@ -694,7 +694,7 @@ class AsyncTimetableAPI(object):
                                 touch_off: datetime | str | None = None,
                                 is_free_fare_zone: bool | None = None,
                                 is_overlap_zone: bool | None = None,
-                                route_types: Iterable[RouteType] | RouteType | None = None
+                                route_types: Iterable[RouteTypeType] | RouteTypeType | None = None
                                 ) -> FareEstimate | None:
         """Returns the estimated fare for the specified journey details.
 
@@ -769,7 +769,7 @@ class AsyncTimetableAPI(object):
     @overload
     async def search(self: Self,
                      search_term: str,
-                     route_types: Iterable[RouteType] | RouteType | None = None,
+                     route_types: Iterable[RouteTypeType] | RouteTypeType | None = None,
                      *,
                      include_outlets: bool | None = None,
                      match_stop_by_locality: bool | None = None,
@@ -795,7 +795,7 @@ class AsyncTimetableAPI(object):
     @overload
     async def search(self: Self,
                      search_term: str,
-                     route_types: Iterable[RouteType] | RouteType | None,
+                     route_types: Iterable[RouteTypeType] | RouteTypeType | None,
                      latitude: float,
                      longitude: float,
                      max_distance: float | None = None,
@@ -808,7 +808,7 @@ class AsyncTimetableAPI(object):
 
     async def search(self: Self,
                      search_term: str,
-                     route_types: Iterable[RouteType] | RouteType | None = None,
+                     route_types: Iterable[RouteTypeType] | RouteTypeType | None = None,
                      latitude: float | None = None,
                      longitude: float | None = None,
                      max_distance: float | None = None,